requirements.txt
python-telegram-bot==21.3
google-generativeai==0.7.1
youtube-api-python==1.10.1